#40: `FromDataframeSampleSizer` should not create empty samples * ensured that sample from non-empty DataFrame is never empty * changed `performRowSizing` to accept optional record count as Long instead of Int

+++ src/main/scala/za/co/absa/spark/partition/sizing/sizer/FromSchemaWithSummariesSizer.scala
@@ -26,7 +26,7 @@ import za.co.absa.spark.partition.sizing.types.{ByteSize, DataTypeSizes}
 @Experimental
 class FromSchemaWithSummariesSizer(implicit dataTypeSizes: DataTypeSizes) extends RecordSizer {
 
-  override def performRowSizing(df: DataFrame, dfRecordCount: Option[Int] = None): ByteSize = {
+  override def performRowSizing(df: DataFrame, dfRecordCount: Option[Long] = None): ByteSize = {
     if(df.isEmpty) 0L
     else {
       val totalCounts: ByteSize = dfRecordCount match {
